03_Airflow/requirements.txt
# Airflow Requirements - ML/MLOps Stack

# Core ML
onnxruntime
onnx
onnxscript
pillow
numpy

# pytorch
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# Data Processing
pandas
scikit-learn

# MLOps
evidently
mlflow

# Database
sqlalchemy
psycopg2-binary

# Cloud Storage
boto3

# HuggingFace
huggingface-hub
# transformers

# API Client
requests

# Testing
pytest
pytest-cov

# Utils
python-dotenv
pydantic
